Fix issue with stale reporter.
Our reporter was getting baked into the saved compiler setup, and then problems
(the SBT class) were lingering between compiles. This reorgs things so that we
create a new 'Setup' on each compile so we have the correct reporter for it.

If the SBT API was not such an incredible pig's breakfast, this might not have
been so easy to fuck up. Anyway, I'm just glad it seems to be working now.